Run python unit test in GitHub workflow
