+++ .github/workflows/build-and-test.yaml
@@ -52,3 +52,45 @@ jobs:
           TF_ON_FLINK_IP: 127.0.0.1
         run: |
           mvn -B test -pl ${{ matrix.modules }}
+  build-and-test-python:
+    runs-on: ubuntu-20.04
+    strategy:
+      fail-fast: false
+      matrix:
+        modules:
+          - dl-on-flink-framework
+          - dl-on-flink-tensorflow
+          - dl-on-flink-tensorflow-2.x
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          submodules: true
+      - name: Set up JDK 8
+        uses: actions/setup-java@v2
+        with:
+          java-version: '8'
+          distribution: 'adopt'
+      - name: Set up python 3.7
+        uses: actions/setup-python@v2
+        with:
+          # Semantic version range syntax or exact version of a Python version
+          python-version: '3.7'
+          # Optional - x64 or x86 architecture, defaults to x64
+          architecture: 'x64'
+      - name: Install Python dependencies
+        run: |
+          pip install -U --user wheel
+      - name: Cache Maven packages
+        uses: actions/cache@v2
+        with:
+          path: ~/.m2
+          key: ${{ runner.os }}-m2-${{ hashFiles('**/pom.xml') }}
+          restore-keys: ${{ runner.os }}-m2
+      - name: build
+        run: |
+          mvn -B install -pl ${{ matrix.modules }} -am -DskipTests -Dfast -Pinstall-py-dep
+      - name: test
+        run: |
+          cd ${{ matrix.modules }}/python
+          pip install -e .
+          python -m unittest discover -v tests
