+++ .github/workflows/prepare-release.yaml
@@ -73,7 +73,50 @@ jobs:
         with:
           name: dl-on-flink-dist
           path: dl-on-flink-dist/target/dl-on-flink-dist-*-bin.tgz
+  build-and-test-python:
+    runs-on: ubuntu-20.04
+    strategy:
+      fail-fast: false
+      matrix:
+        modules:
+          - dl-on-flink-framework
+          - dl-on-flink-tensorflow
+          - dl-on-flink-tensorflow-2.x
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          submodules: true
+      - name: Set up JDK 8
+        uses: actions/setup-java@v2
+        with:
+          java-version: '8'
+          distribution: 'adopt'
+      - name: Set up python 3.7
+        uses: actions/setup-python@v2
+        with:
+          # Semantic version range syntax or exact version of a Python version
+          python-version: '3.7'
+          # Optional - x64 or x86 architecture, defaults to x64
+          architecture: 'x64'
+      - name: Install Python dependencies
+        run: |
+          pip install -U --user wheel
+      - name: Cache Maven packages
+        uses: actions/cache@v2
+        with:
+          path: ~/.m2
+          key: ${{ runner.os }}-m2-${{ hashFiles('**/pom.xml') }}
+          restore-keys: ${{ runner.os }}-m2
+      - name: build
+        run: |
+          mvn -B install -pl ${{ matrix.modules }} -am -DskipTests -Dfast -Pinstall-py-dep
+      - name: test
+        run: |
+          cd ${{ matrix.modules }}/python
+          pip install -e .
+          python -m unittest discover -v tests
   build-linux-wheel:
+    needs: build-and-test-python
     name: Build python${{ matrix.python-version }} ${{ matrix.package }} wheels on ubuntu-20.04
     runs-on: ubuntu-20.04
     strategy:
@@ -102,6 +145,7 @@ jobs:
           name: dl-on-flink-wheels
           path: ./wheelhouse/*.whl
   build-macos-wheel:
+    needs: build-and-test-python
     name: Build python${{ matrix.python-version }} ${{ matrix.package }} wheels on macOS-10.15
     runs-on: macos-10.15
     strategy:
