+++ dl-on-flink-framework/python/tests/test_sysconfig.py
@@ -1,0 +1,27 @@
+#  Licensed to the Apache Software Foundation (ASF) under one
+#  or more contributor license agreements.  See the NOTICE file
+#  distributed with this work for additional information
+#  regarding copyright ownership.  The ASF licenses this file
+#  to you under the Apache License, Version 2.0 (the
+#  "License"); you may not use this file except in compliance
+#  with the License.  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+import os.path
+import unittest
+
+from dl_on_flink_framework import sysconfig
+
+
+class TestSysconfig(unittest.TestCase):
+
+    def test_get_include(self):
+        expected = os.path.join(os.path.dirname(os.path.dirname(__file__)),
+                                "dl_on_flink_framework", "include")
+        self.assertEqual(expected, sysconfig.get_include())
